requirements.txt
Flask==2.0.1
Flask-CORS==4.0.0
python-dotenv==1.0.0
gunicorn==20.1.0
Werkzeug==2.0.3
SQLAlchemy==1.4.23
psutil==5.9.0